+++ cmd/list.go
@@ -1,7 +1,9 @@
 package cmd
 
 import (
+	"encoding/json"
 	"fmt"
+	"os"
 	"sort"
 	"time"
 
@@ -11,6 +13,20 @@ import (
 
 	"github.com/spf13/cobra"
 )
+
+func textDump(lines []string) {
+	sort.Strings(lines)
+	for _, line := range lines {
+		common.Log("%s", line)
+	}
+}
+
+func jsonDump(entries map[string]interface{}) {
+	body, err := json.MarshalIndent(entries, "", "  ")
+	if err == nil {
+		fmt.Fprintln(os.Stdout, string(body))
+	}
+}
 
 var listCmd = &cobra.Command{
 	Use:   "list",
@@ -23,8 +39,13 @@ in human readable form.`,
 			pretty.Exit(1, "No environments available.")
 		}
 		lines := make([]string, 0, len(templates))
-		common.Log("%-25s  %-25s  %-16s  %s", "Last used", "Last cloned", "Environment", "Leased duration")
+		entries := make(map[string]interface{})
+		if !jsonFlag {
+			common.Log("%-25s  %-25s  %-16s  %5s  %s", "Last used", "Last cloned", "Environment", "Plan?", "Leased duration")
+		}
 		for _, template := range templates {
+			details := make(map[string]interface{})
+			entries[template] = details
 			cloned := "N/A"
 			used := cloned
 			when, err := conda.LastUsed(conda.TemplateFrom(template))
@@ -35,15 +56,28 @@ in human readable form.`,
 			if err == nil {
 				used = when.Format(time.RFC3339)
 			}
-			lines = append(lines, fmt.Sprintf("%-25s  %-25s  %-16s  %q %s", used, cloned, template, conda.WhoLeased(template), conda.LeaseExpires(template)))
+			details["name"] = template
+			details["used"] = used
+			details["cloned"] = cloned
+			details["leased"] = conda.WhoLeased(template)
+			details["expires"] = conda.LeaseExpires(template)
+			details["base"] = conda.TemplateFrom(template)
+			details["live"] = conda.LiveFrom(template)
+			planfile, plan := conda.InstallationPlan(template)
+			lines = append(lines, fmt.Sprintf("%-25s  %-25s  %-16s  %5v  %q %s", used, cloned, template, plan, conda.WhoLeased(template), conda.LeaseExpires(template)))
+			if plan {
+				details["plan"] = planfile
+			}
 		}
-		sort.Strings(lines)
-		for _, line := range lines {
-			common.Log("%s", line)
+		if jsonFlag {
+			jsonDump(entries)
+		} else {
+			textDump(lines)
 		}
 	},
 }
 
 func init() {
 	envCmd.AddCommand(listCmd)
+	listCmd.Flags().BoolVarP(&jsonFlag, "json", "j", false, "Output in JSON format.")
 }